backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Data Validation
pydantic==2.10.0
pydantic-settings==2.6.0

# Database
supabase==2.10.0
asyncpg==0.30.0

# AI & Agents
openai>=1.99.5
opik==1.9.87

# Optimization
pulp==2.9.0

# Cryptography
cryptography==44.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# HTTP Client
httpx==0.27.2

# Testing
pytest==8.3.0
pytest-asyncio==0.25.0
pytest-cov==6.0.0
hypothesis==6.122.0

# Code Quality
black==24.10.0
ruff==0.8.0
mypy==1.13.0

# Environment
python-dotenv==1.0.1
